+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+### Changed
+- Allow parsing of types that do not implement `ToTokens`
+
 ## [0.10.3] - 2023-11-23
 ### Added
 - `impl SpanRanged for Range<impl SpanRanged>`