+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+<!-- ## [Unreleased] -->
+## [0.10.4] - 2023-11-24
 ### Changed
 - Allow parsing of types that do not implement `ToTokens`
 
@@ -84,7 +85,8 @@ Only documentation changes.
 ## [v0.1.0] 
 **Initial Release**
 
-[unreleased]: https://github.com/ModProg/manyhow/compare/v0.10.3...HEAD
+[unreleased]: https://github.com/ModProg/manyhow/compare/v0.10.4...HEAD
+[0.10.4]: https://github.com/ModProg/manyhow/compare/v0.10.3...v0.10.4
 [0.10.3]: https://github.com/ModProg/manyhow/compare/v0.10.2...v0.10.3
 [0.10.2]: https://github.com/ModProg/manyhow/compare/v0.10.1...v0.10.2
 [0.10.1]: https://github.com/ModProg/manyhow/compare/v0.10.0...v0.10.1